api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo[srv]==4.6.1
passlib==1.7.4
bcrypt==3.2.2
pyjwt==2.8.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.0.3
slowapi==0.1.9
